Add files via upload
This program uses Binary Search Tree's. 

It first searches for any email (through regular expressions) and then searches that email against a set of given emails (16k in our example) in the BST for matches. If there is a match, it returns the term that was searched, the file it was found in, and the entire line it was found on.